refactor: AnimatedText style 분리
