+++ src/components/AnimatedText/styled.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const AnimatedTextContainer = styled.div`
+  position: relative;
+  width: 100%;
+  height: 100%;
+`;
